src/requirements.txt
Django==1.10.5
psycopg2==2.6.2
uWSGI==2.0.14
requirements==2.13.0
imdbpie==4.2.0
djangorestframework==3.6.2
django-extensions==1.7.7